requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
audioread==3.0.1
    # via librosa
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via streamlit
certifi==2025.1.31
    # via requests
cffi==1.17.1
    # via soundfile
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   nltk
    #   streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via librosa
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.57.0
    # via matplotlib
fsspec==2025.3.2
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
huggingface-hub==0.30.2
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.4.2
    # via
    #   librosa
    #   nltk
    #   pynndescent
    #   scikit-learn
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
kiwisolver==1.4.8
    # via matplotlib
lazy-loader==0.4
    # via librosa
librosa==0.11.0
    # via -r requirements.in
llvmlite==0.44.0
    # via
    #   numba
    #   pynndescent
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.1
    # via
    #   -r requirements.in
    #   seaborn
mpmath==1.3.0
    # via sympy
msgpack==1.1.0
    # via librosa
narwhals==1.35.0
    # via altair
networkx==3.4.2
    # via torch
nltk==3.9.1
    # via -r requirements.in
numba==0.61.2
    # via
    #   librosa
    #   pynndescent
    #   umap-learn
numpy==2.2.5
    # via
    #   -r requirements.in
    #   contourpy
    #   librosa
    #   matplotlib
    #   numba
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   seaborn
    #   soundfile
    #   soxr
    #   streamlit
    #   transformers
    #   umap-learn
packaging==24.2
    # via
    #   altair
    #   huggingface-hub
    #   lazy-loader
    #   matplotlib
    #   pooch
    #   streamlit
    #   transformers
pandas==2.2.3
    # via
    #   seaborn
    #   streamlit
pillow==11.2.1
    # via
    #   matplotlib
    #   streamlit
platformdirs==4.3.7
    # via pooch
pooch==1.8.2
    # via librosa
protobuf==5.29.4
    # via streamlit
pyarrow==19.0.1
    # via streamlit
pycparser==2.22
    # via cffi
pydeck==0.9.1
    # via streamlit
pydub==0.25.1
    # via -r requirements.in
pynndescent==0.5.13
    # via umap-learn
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   nltk
    #   transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   pooch
    #   streamlit
    #   transformers
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via
    #   -r requirements.in
    #   librosa
    #   pynndescent
    #   umap-learn
scipy==1.15.2
    # via
    #   -r requirements.in
    #   librosa
    #   pynndescent
    #   scikit-learn
    #   umap-learn
seaborn==0.13.2
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
soundfile==0.13.1
    # via
    #   -r requirements.in
    #   librosa
soxr==0.5.0.post1
    # via librosa
streamlit==1.44.1
    # via -r requirements.in
sympy==1.13.1
    # via torch
tenacity==9.1.2
    # via streamlit
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.6.0
    # via -r requirements.in
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   nltk
    #   transformers
    #   umap-learn
transformers==4.51.3
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   altair
    #   huggingface-hub
    #   librosa
    #   referencing
    #   streamlit
    #   torch
tzdata==2025.2
    # via pandas
umap-learn==0.5.7
    # via -r requirements.in
urllib3==2.4.0
    # via requests
watchdog==6.0.0
    # via streamlit
